Create Category Edit Painting

+++ app/Repositories/PaintingRepository.php
@@ -5,6 +5,8 @@ namespace App\Repositories;
 use App\Http\CheckNullableInterface;
 use App\Http\CheckNullableParams;
 use App\Models\Painting;
+use App\Models\PaintingImage;
+use Illuminate\Support\Facades\Storage;
 use RuntimeException;
 
 
@@ -26,7 +28,9 @@ class PaintingRepository implements CheckNullableInterface
                 }
             }
         }else{
-            throw new RuntimeException($value . ' is required ! Now is null');
+            if($check){
+                throw new RuntimeException($value . ' is required ! Now is null');
+            }
         }
     }
 
@@ -50,26 +54,11 @@ class PaintingRepository implements CheckNullableInterface
         return $this->painting->with('images')->find($paintingId);
     }
 
-    public function storePainting($name, $name_en, $price, $description, $description_en, $year, $width, $height, $long, $sold_status)
+    public function storePainting($painting)
     {
-        $this->checkNullable([$name, $name_en, $price, $description, $description_en, $year, $width, $height, $long, $sold_status]);
+        $this->checkNullable($painting);
 
-        if($sold_status){
-            $sold_status = 1;
-        }
-
-        $this->painting->create([
-            'name' => $name,
-            'name_en' => $name_en,
-            'price' => $price,
-            'description' => $description,
-            'description_en' => $description_en,
-            'year' => $year,
-            'width' => $width,
-            'height' => $height,
-            'long' => $long,
-            'sold_status' => $sold_status,
-        ]);
+        $this->painting->create($painting);
 
         return $this->painting->all()->last()->id;
     }
@@ -85,7 +74,19 @@ class PaintingRepository implements CheckNullableInterface
     {
         $this->checkNullable($paintingId);
 
-        $this->painting->destroy($paintingId);
+        $images = PaintingImage::where('painting_id', $paintingId)->get() ?? null;
+
+        if($images !== null){
+            foreach ($images as $image){
+                if(Storage::disk('public')->exists('/images/' . $image->painting_id . '_' . $image->hash_id)){
+                    Storage::disk('public')->delete('/images/' . $image->painting_id . '_' . $image->hash_id);
+                }
+            }
+            $this->painting->destroy($paintingId);
+        }
+
+
+
     }
 
 }
